requirements.txt
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
importlib_metadata==8.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
PyJWT==2.10.1
python-arango==8.2.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
setuptools==80.9.0
urllib3==2.6.1
zipp==3.23.0
